Feat: 암호 풀기 기능 개발

See merge request s11-final/S11P31A404!31

+++ backend/src/main/java/com/ssafy/fiftyninesec/solution/entity/Prize.java
@@ -15,8 +15,8 @@ import lombok.*;
 public class Prize {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer prizeId;
-    private Integer roomId;
+    private Long prizeId;
+    private Long roomId;
     private String prizeType;
     private String prizeName;
     private Integer winnerCount;
